Split up trivia module and add hangman
